Refactor check into schema extensions class
